+++ lib/rubycanusellm/tool_call.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module RubyCanUseLLM
+  class ToolCall
+    attr_reader :id, :name, :arguments
+
+    def initialize(id:, name:, arguments:)
+      @id = id
+      @name = name
+      @arguments = arguments
+    end
+  end
+end